Begin porting 2.0.0 - switch to double currency

+++ fabric/src/main/java/the_fireplace/grandeconomy/fabric/api/event/BalanceChangeEvent.java
@@ -12,5 +12,5 @@ public interface BalanceChangeEvent {
                 event.onBalanceChanged(accountId, oldBalance, newBalance);
         });
 
-    void onBalanceChanged(UUID accountId, long oldBalance, long newBalance);
+    void onBalanceChanged(UUID accountId, double oldBalance, double newBalance);
 }
