+++ forge/src/main/java/the_fireplace/grandeconomy/forge/econhandlers/fe/ForgeEssentialsEconHandler.java
@@ -2,34 +2,34 @@ package the_fireplace.grandeconomy.forge.econhandlers.fe;
 
 /*public class ForgeEssentialsEconHandler implements IEconHandler {
     @Override
-    public long getBalance(UUID uuid, Boolean isPlayer) {
+    public double getBalance(UUID uuid, Boolean isPlayer) {
         return APIRegistry.economy.getWallet(UserIdent.get(uuid)).get();
     }
 
     @Override
-    public boolean addToBalance(UUID uuid, long amount, Boolean isPlayer) {
+    public boolean addToBalance(UUID uuid, double amount, Boolean isPlayer) {
         APIRegistry.economy.getWallet(UserIdent.get(uuid)).add(amount);
         return true;
     }
 
     @Override
-    public boolean takeFromBalance(UUID uuid, long amount, Boolean isPlayer) {
+    public boolean takeFromBalance(UUID uuid, double amount, Boolean isPlayer) {
         return APIRegistry.economy.getWallet(UserIdent.get(uuid)).withdraw(amount);
     }
 
     @Override
-    public boolean setBalance(UUID uuid, long amount, Boolean isPlayer) {
+    public boolean setBalance(UUID uuid, double amount, Boolean isPlayer) {
         APIRegistry.economy.getWallet(UserIdent.get(uuid)).set(amount);
         return true;
     }
 
     @Override
-    public String getCurrencyName(long amount) {
+    public String getCurrencyName(double amount) {
         return APIRegistry.economy.currency(amount);
     }
 
     @Override
-    public String toString(long amount) {
+    public String toString(double amount) {
         return APIRegistry.economy.toString(amount);
     }
